backend/requirements (copy).txt

absl-py
arabic-reshaper
asgiref
astunparse
cachetools
certifi
charset-normalizer
dj-database-url
Django
django-cors-headers
django-heroku
djangorestframework
et-xmlfile
flatbuffers
future
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
gunicorn
h5py
html5lib
idna
importlib-metadata
joblib
jsonfield
keras
libclang
Markdown
mysqlclient
numpy
oauthlib
openpyxl
opt-einsum
pandas
Pillow
protobuf
# psycopg2-binary==2.9.1
pyasn1
pyasn1-modules
python-bidi
python-dateutil
pytz
reportlab
requests
requests-oauthlib
rsa
scikit-learn
scipy
six
scikit-learn
sqlparse
tensorflow
termcolor
tf-estimator-nightly
threadpoolctl
typing_extensions
urllib3
webencodings
Werkzeug
whitenoise
wincertstore
wrapt
xhtml2pdf
zipp
nltk
gensim

absl-py==1.0.0
arabic-reshaper==2.1.3
asgiref==3.3.4
astunparse==1.6.3
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
flatbuffers==2.0
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.0
joblib==1.1.0
jsonfield==3.1.0
keras==2.8.0
libclang==13.0.0
Markdown==3.3.6
mysqlclient==2.1.0
numpy==1.21.5
oauthlib==3.2.0
openpyxl==3.0.9
opt-einsum==3.3.0
pandas==1.3.5
Pillow==8.2.0
protobuf==3.19.4
# psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
scikit-learn
sqlparse==0.4.1
tensorflow==2.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
typing_extensions==4.1.0
urllib3==1.26.8
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.13.3
xhtml2pdf==0.2.5
zipp==3.7.0
nltk
gensim==3



# Django-4.2.3 Markdown-3.4.3 MarkupSafe-2.1.3 Pillow-10.0.0 Werkzeug-2.3.6 absl-py-1.4.0 arabic-reshaper-3.0.0 asgiref-3.7.2 asn1crypto-1.5.1 astunparse-1.6.3 cachetools-5.3.1 certifi-2023.5.7 cffi-1.15.1 charset-normalizer-3.2.0 click-8.1.5 contourpy-1.1.0 cryptography-41.0.2 cssselect2-0.7.0 cycler-0.11.0 dj-database-url-2.0.0 django-cors-headers-4.2.0 django-heroku-0.3.1 djangorestframework-3.14.0 et-xmlfile-1.1.0 flatbuffers-23.5.26 fonttools-4.41.0 future-0.18.3 gast-0.4.0 gensim-4.3.1 google-auth-2.22.0 google-auth-oauthlib-1.0.0 google-pasta-0.2.0 grpcio-1.56.0 gunicorn-20.1.0 h5py-3.9.0 html5lib-1.1 idna-3.4 importlib-metadata-6.8.0 importlib-resources-6.0.0 joblib-1.3.1 jsonfield-3.1.0 keras-2.13.1 kiwisolver-1.4.4 libclang-16.0.0 lxml-4.9.3 matplotlib-3.7.2 mysql-connector-python-8.0.26 mysqlclient-2.1.0 nltk-3.8.1 numpy-1.24.3 oauthlib-3.2.2 openpyxl-3.1.2 opt-einsum-3.3.0 oscrypto-1.3.0 packaging-23.1 pandas-2.0.3 protobuf-4.23.4 psycopg2-2.9.6 pyHanko-0.19.0 pyasn1-0.5.0 pyasn1-modules-0.3.0 pycparser-2.21 pyhanko-certvalidator-0.23.0 pyparsing-3.0.9 pypdf-3.12.1 pypng-0.20220715.0 python-bidi-0.4.2 python-dateutil-2.8.2 pytz-2023.3 pyyaml-6.0 qrcode-7.4.2 regex-2023.6.3 reportlab-3.6.13 requests-2.31.0 requests-oauthlib-1.3.1 rsa-4.9 scikit-learn-1.3.0 scipy-1.11.1 seaborn-0.12.2 six-1.16.0 smart-open-6.3.0 sqlparse-0.4.4 svglib-1.5.1 tensorboard-2.13.0 tensorboard-data-server-0.7.1 tensorflow-2.13.0 tensorflow-estimator-2.13.0 tensorflow-io-gcs-filesystem-0.32.0 termcolor-2.3.0 tf-estimator-nightly-2.14.0.dev2023071408 threadpoolctl-3.2.0 tinycss2-1.2.1 tqdm-4.65.0 typing_extensions-4.5.0 tzdata-2023.3 tzlocal-5.0.1 uritools-4.0.1 urllib3-1.26.16 webencodings-0.5.1 wheel-0.40.0 whitenoise-6.5.0 wincertstore-0.2 wrapt-1.15.0 xhtml2pdf-0.2.11 zipp-3.16.1

Django==4.2.3
Markdown==3.4.3
MarkupSafe==2.1.3
Pillow==10.0.0
Werkzeug==2.3.6
absl-py==1.4.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
contourpy==1.1.0
cryptography==41.0.2
cssselect2==0.7.0
cycler==0.11.0
dj-database-url==2.0.0
django-cors-headers==4.2.0
# django-heroku==0.3.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
flatbuffers==23.5.26
fonttools==4.41.0
future==0.18.3
gast==0.4.0
gensim==4.3.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
gunicorn==20.1.0
h5py==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
joblib==1.3.1
jsonfield==3.1.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.3
matplotlib==3.7.2
mysql-connector-python==8.0.26
mysqlclient==2.1.0
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.1
pandas==2.0.3
protobuf==4.23.4
psycopg2==2.9.6
pyHanko==0.19.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyhanko-certvalidator==0.23.0
pyparsing==3.0.9
pypdf==3.12.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0
qrcode==7.4.2
regex==2023.6.3
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smart-open==6.3.0
sqlparse==0.4.4
svglib==1.5.1
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tf-estimator-nightly==2.14.0.dev2023071408
threadpoolctl==3.2.0
tinycss2==1.2.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.16
webencodings==0.5.1
wheel==0.40.0
whitenoise==6.5.0
wincertstore==0.2
wrapt==1.15.0
xhtml2pdf==0.2.11
zipp==3.16.1
xlwt==1.3.0
